Remove small dialog
Removed the small dialog displayed after the Export Layers button is clicked. The progress bar and the Stop button are now displayed directly in the main dialog.
The file format dialog should now appear in front of the main dialog - at the very least, it works on Ubuntu. It doesn't work on Windows, however. You'll need to switch to the file format dialog manually in the taskbar.