Use the new presenter for directory choosers
To remember the image list and the last export path of each image across GIMP sessions, a new setting is defined.
This only applies to the interactive run mode. In non-interactive mode, it would be counter-intuitive, since the user expects that the output directory specified as a parameter will actually be used.