pgitemdata: Improve determining file extension from item names
The `pgfileformats` module is now used to get/set the file extension in item names. This is useful when the file extension contains a period (e.g. "xcf.bz2") and also resolves issues where such extension is expected to be returned instead of always returning the substring after the last period.