Export name preview: Allow to modify tags in items
One can now modify tags of a layer by right-clicking on the layer in the preview and checking/unchecking available tags. The layer name gets updated as well as the visibility of the tag icon.
Also removed multiple representations of the suporrted tags (e.g. "background", "bg") since tag editing in the GUI is simpler and less error-prone than typing tags in the layer name.