ItemData: Optimize performance by reducing PDB calls
It turns out that processing hundreds of layers can be very resource-intensive - as it turns out, this is due to a high number of calls to the GIMP PDB. By reducing the number of redundant PDB calls, we reduced the performance overhead of instantiating LayerData from roughly 1.8 to roughly 0.5 seconds for 900+ layers. The optimization also applies to other ItemData types, though probably to a lesser degree.