pgsetting: Fix resetting to default value for empty containers
If the setting has a container as the default value (dict, list, ...), resetting to its default value works properly if the default value is an empty container.
If the default value contains other containers or class instances, these may not be copied and the default value may still be modified. In other words, deep copy is not performed.