Rename "advanced" settings to "more" settings
The settings weren't really "advanced" in the first place, they're simply less commonly used (or so I think).
Also, replaced "Advanced settings" expander with a "Show more settings" check button since the preview that gets displayed is not placed below the expander.
Also, moved the check button next to the dialog buttons. The default dialog height was slightly decreased to account for the vertical space saved by removing the expander.