Access layers by their IDs instead of their names
This fixes a bug where an item would be moved at the end of the name preview if the user modified tags for the item. This change is also beneficial because, unlike layer names, IDs never change and thus we avoid subtle bugs and ugly workarounds for tag editing in the name preview.