Modify behavior of "Crop mode" setting and resizing layers
If "Use image size" is disabled, the layer is now always resized to its own size rather than the combined size of the layer and specially treated tagged layers. An additional setting will be created in the future to resize to the combined size.
The "Crop mode" setting has been split to "Crop to background" and "Crop to foreground". It is now possible to autocrop to the layer and the tagged layers at the same time.