gui_preview_image: Properly scale for custom operations
Since any procedure can now be applied, such as paintbrush or
downscaling, the processing needs to be performed on the original size
of the layers and then downscaled at the end of the processing. If no
custom operation is present, downscaling is (as is the case now)
performed prior to processing, saving performance for large layers.